+++ backend/src/docs/asciidoc/car.adoc
@@ -40,3 +40,30 @@ include::{snippets}/car-controller-test/delete_car/http-request.adoc[]
 === Response
 
 include::{snippets}/car-controller-test/delete_car/http-response.adoc[]
+
+
+== 회원이 등록한 차량을 조회한다
+
+=== Request
+
+include::{snippets}/car-controller-test/find_member_car/request-headers.adoc[]
+include::{snippets}/car-controller-test/find_member_car/http-request.adoc[]
+
+=== Response
+
+include::{snippets}/car-controller-test/find_member_car/response-fields.adoc[]
+include::{snippets}/car-controller-test/find_member_car/http-response.adoc[]
+
+== 회원이 차량을 등록한다
+
+=== Request
+
+include::{snippets}/car-controller-test/add_member_car/request-headers.adoc[]
+include::{snippets}/car-controller-test/add_member_car/path-parameters.adoc[]
+include::{snippets}/car-controller-test/add_member_car/request-fields.adoc[]
+include::{snippets}/car-controller-test/add_member_car/http-request.adoc[]
+
+=== Response
+
+include::{snippets}/car-controller-test/add_member_car/response-fields.adoc[]
+include::{snippets}/car-controller-test/add_member_car/http-response.adoc[]
