+++ backend/src/test/java/com/carffeine/carffeine/car/controller/CarControllerTest.java
@@ -6,6 +6,8 @@ import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
 import com.carffeine.carffeine.car.service.dto.CarsRequest;
 import com.carffeine.carffeine.helper.MockBeanInjection;
+import com.carffeine.carffeine.member.domain.MemberCar;
+import com.carffeine.carffeine.member.fixture.MemberFixture;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -133,5 +135,60 @@ class CarControllerTest extends MockBeanInjection {
                         requestHeaders(headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")),
                         pathParameters(parameterWithName("carId").description("차량 id"))
                 ));
+    }
+
+    @Test
+    void 회원의_차량을_등록한다() throws Exception {
+        // given
+        Car car = createCar();
+        CarRequest carRequest = new CarRequest("아이오닉5", "2022-A");
+
+        // when
+        when(carService.addMemberCar(any(), any(), any())).thenReturn(car);
+
+        // then
+        mockMvc.perform(post("/members/{memberId}/cars", 1L)
+                        .header(org.springframework.http.HttpHeaders.AUTHORIZATION, "Bearer token~~")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(carRequest))
+                ).andExpect(status().isCreated())
+                .andDo(customDocument("add_member_car",
+                        requestHeaders(headerWithName(org.springframework.http.HttpHeaders.AUTHORIZATION).description("인증 토큰")),
+                        pathParameters(parameterWithName("memberId").description("Member id")),
+                        requestFields(
+                                fieldWithPath("name").type(JsonFieldType.STRING).description("차량 이름"),
+                                fieldWithPath("vintage").type(JsonFieldType.STRING).description("차량 연식")
+                        ),
+                        responseFields(
+                                fieldWithPath("carId").type(JsonFieldType.NUMBER).description("차량 id"),
+                                fieldWithPath("name").type(JsonFieldType.STRING).description("차량 이름"),
+                                fieldWithPath("vintage").type(JsonFieldType.STRING).description("차량 연식")
+                        )
+                ));
+    }
+
+
+    @Test
+    void 회원이_등록한_차량을_조회한다() throws Exception {
+        // given
+        MemberCar memberCar = new MemberCar(MemberFixture.일반_회원, createCar());
+
+        // when
+        when(carService.findMemberCar(any())).thenReturn(memberCar);
+
+        // then
+        mockMvc.perform(get("/members/me")
+                        .header(org.springframework.http.HttpHeaders.AUTHORIZATION, "Bearer token~~")
+                ).andExpect(status().isOk())
+                .andDo(customDocument("find_member_car",
+                                requestHeaders(headerWithName(org.springframework.http.HttpHeaders.AUTHORIZATION).description("인증 토큰")),
+                                responseFields(
+                                        fieldWithPath("memberId").type(JsonFieldType.NUMBER).description("유저 id"),
+                                        fieldWithPath("car.carId").type(JsonFieldType.NUMBER).description("차량 id"),
+                                        fieldWithPath("car.name").type(JsonFieldType.STRING).description("차량 이름"),
+                                        fieldWithPath("car.vintage").type(JsonFieldType.STRING).description("차량 연식")
+                                )
+                        )
+                );
     }
 }
